pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]  # Build tools needed
build-backend = "setuptools.build_meta"

[project]
name = "ailibrary"
version = "0.1.2"
description = "AI Library REST API for any Python 3.9+ application."
requires-python = ">=3.9"
readme = "README.md"
authors = [
    {name = "Arani Chaudhuri", email = "arani@ailibrary.ai"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
license = {text = ""} #?
dynamic = ["dependencies"]

[project.optional-dependencies]
test = [
    "pytest>=8.3.4",
    "pytest-cov",
    "pytest-mock",
    "pytest-order",
    # "pytest-asyncio",  # if you need async testing
    "python-dotenv>=1.0.1"
]
dev = [
    "black",
    "flake8",
    "mypy"
]

[tool.setuptools.dynamic]
dependencies = {file = ["requirements.txt"]}

[tool.setuptools.packages.find]
where = ["src"]  # Tell setuptools to look for packages in the 'src' folder
